all: fix build
Add a wasm build constraint to a WebAssembly-specific command.

Encountering an empty go.mod file can be confusing for humans
and for builders. Remove it, it's probably not needed anymore.

Lower the go directive to 1.22 in testtiming so that a go1.22
toolchain can test it without needing to upgrade.

Change-Id: I6550f67815d7a9a5fb70db57b17e576526fa8c8b
Reviewed-on: https://go-review.googlesource.com/c/scratch/+/610455
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>